final, documentation yet to be documentation

+++ cart.js
@@ -21,24 +21,17 @@ let cartSubTotal=(new_order)=>{
     console.log(new_order);
     let subTotal=0;
      let quantitySoFar=new_order.counter;
-     console.log(quantitySoFar+'quan')
-
     let objName = 'item';
     for(let i=1; i <= +quantitySoFar; i++ ){
-        let thisObjName= objName + quantitySoFar;
+        let thisObjName= objName + i;
         let price=+new_order[thisObjName].product_unit_price.match(/\d+/g)*(+new_order[thisObjName].quantity);
-        console.log(i);
-       
-        subTotal=subTotal+(price); 
-        
+        console.log(i);   
+        subTotal=subTotal+(price);    
     }
   
     return subTotal;       
 }
     
-  
-  
-
 
 //Delete Handler !!!!!!
 
@@ -73,20 +66,19 @@ function remove(itemName){
    
     
 }
- 
 function changeQty(thisid){
   let theid= document.getElementById(thisid).getAttribute('id');
  let value = document.getElementById(theid).value;
- console.log(theid + value);
+ if (+value > 0 && +value <= 10){
  let jsonString=localStorage.getItem("new_order");
  let new_order = JSON.parse(jsonString);
  new_order['item'+ theid].quantity=+value;
  new_order.subTotal= cartSubTotal(new_order);
  new_order.discounted = cartSubTotal(new_order)/2       
-   
+   console.log(value);
  localStorage.clear();
  localStorage.setItem("new_order", JSON.stringify(new_order));
- 
+ }
  createCartObjects()
 }
 
@@ -114,7 +106,7 @@ let createCartObjects=()=>{
         let displayCoupon = document.getElementById('coupon_code');
         
 
-                    if (new_order && new_order.counter>=1){
+                    if (new_order){
                         for(let i=1; i<= new_order.counter; i++ ){
                             
                             itemNum = 'item'+ i;
@@ -147,6 +139,8 @@ let createCartObjects=()=>{
                         displaysTotal.innerText=new_order.subTotal;
                         displayEst.innerText=new_order.subTotal;
                         displayCoupon.placeholder=new_order.discount_coupon;
+                        document.getElementById('s_counter1').innerHTML=new_order.counter;
+                        document.getElementById('s_counter').innerText=new_order.counter;
 
 
                     }
@@ -164,35 +158,36 @@ function htmlFormat(c){
 </div>
 <div class="description">
     <h4 id="describe${c}" ></h4><br>
-    <!-- <h6>Style</h6>: 2D023 |  -->
     <h6 id="color${c}"></h6>
     <h6 id="size${c}">Size</h6> <br>
+    <div class="updater">
     <p><input class="multiplier" id=${c} type="number" min="1" step="1" max="10")">
         <button onclick="changeQty(${c})" >Update</button></p>
-    </select>
+    
     <h6><div>Unit price</div><br><div id="unit_price${c}">120</div></h6> &nbsp;&nbsp;&nbsp;
     <h6><div>Total Price</div> <br><div id="total_price${c}">$120</div></h6>
+    </div>
     <div class="right">
         <div class="chckbx">
             <input type="checkbox"> Free gift package?
         </div>
-        <hr>
+        <hr><br>
         <div class="radiobtn">
-            <input type="radio" checked> <b>Ship to Me</b> 
+            <input type="radio" checked name = "radio"> <span>Ship to Me</span> <br>
+            <input type="radio" name = "radio"> <span>In-Store Pickup </span><br>
+            <input type="radio" name = "radio"> <span>Curbside Pickup</span>
         </div>
-        <div class="radiobtn">
-            <input type="radio"> In-Store Pickup
-        </div>
-        <div class="radiobtn">
-            <input type="radio"> Curbside Pickup
-        </div>
+    
+       
         <div class="remedit">
         <div> <a id="item${c}" onclick="remove(item${c})" href="">Remove</a></div></div>
     </div>
+   
 
 </div>
 
 </div>
+
 `;
 return text;
 }
